simplify hello.cwl
The first encounter with the cwl for executing something needs to be as simple as ... donno ... something really simple.  The traditional "Hello world!" for instance does not need any input. It is a mere generator.

When executing the cwltool, I was asked if I wanted to to run #echocmd or #main. So, if #main is obsolete, which is good, then I strongly suggest not to have it with hello. With a bit of an imagination the CommandLineTool is a special kind of workflow, i.e. one that cannot nest or interconnect other workflows but wraps a tool. Then it is just fine to execute it direcly.

The input line should not be required for a command line. A missing such should just be interpreted as 'no inputs'.

I also removed the "stdout" line to make it more interactive, i.e. just like the original.